requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
groq
python-dotenv